refactor: move enableAssetFallback option to uni-builder
