+++ e2e/cases/css/less-exclude/index.test.ts
@@ -8,12 +8,19 @@ test('should exclude specified less file', async () => {
     entry: { index: path.resolve(__dirname, './src/index.js') },
     rsbuildConfig: {
       tools: {
-        less: (opts, { addExcludes }) => {
+        less: (_, { addExcludes }) => {
           addExcludes([/b\.less$/]);
         },
-      },
-      output: {
-        enableAssetFallback: true,
+        rspack: {
+          module: {
+            rules: [
+              {
+                test: /b\.less$/,
+                type: 'asset/resource',
+              },
+            ],
+          },
+        },
       },
     },
   });
