+++ e2e/cases/css/sass-exclude/index.test.ts
@@ -8,12 +8,19 @@ test('should exclude specified scss file', async () => {
     entry: { index: path.resolve(__dirname, './src/index.js') },
     rsbuildConfig: {
       tools: {
-        sass: (opts, { addExcludes }) => {
+        sass: (_, { addExcludes }) => {
           addExcludes([/b\.scss$/]);
         },
-      },
-      output: {
-        enableAssetFallback: true,
+        rspack: {
+          module: {
+            rules: [
+              {
+                test: /b\.scss$/,
+                type: 'asset/resource',
+              },
+            ],
+          },
+        },
       },
     },
   });
