+++ packages/compat/uni-builder/src/types.ts
@@ -79,6 +79,11 @@ export type UniBuilderExtraConfig = {
      * Whether to treat all .css files in the source directory as CSS Modules.
      */
     disableCssModuleExtension?: boolean;
+    /**
+     * If this option is enabled, all unrecognized files will be emitted to the dist directory.
+     * Otherwise, an exception will be thrown.
+     */
+    enableAssetFallback?: boolean;
   };
   html?: {
     /**
