+++ packages/compat/webpack/src/shared/plugin.ts
@@ -38,5 +38,4 @@ export const applyDefaultPlugins = (plugins: Plugins) =>
     plugins.performance(),
     plugins.networkPerformance(),
     plugins.preloadOrPrefetch(),
-    import('../plugins/fallback').then((m) => m.pluginFallback()), // fallback should be the last plugin
   ]);
