+++ packages/document/docs/en/config/options/output.mdx
@@ -72,12 +72,6 @@ import DisableFilenameHash from '@en/shared/config/output/disableFilenameHash.md
 
 <DisableFilenameHash />
 
-## output.enableAssetFallback
-
-import EnableAssetFallback from '@en/shared/config/output/enableAssetFallback.md';
-
-<EnableAssetFallback />
-
 ## output.enableCssModuleTSDeclaration
 
 import EnableCssModuleTSDeclaration from '@en/shared/config/output/enableCssModuleTSDeclaration.md';
