+++ packages/document/docs/zh/config/options/output.mdx
@@ -72,12 +72,6 @@ import DisableFilenameHash from '@zh/shared/config/output/disableFilenameHash.md
 
 <DisableFilenameHash />
 
-## output.enableAssetFallback
-
-import EnableAssetFallback from '@zh/shared/config/output/enableAssetFallback.md';
-
-<EnableAssetFallback />
-
 ## output.enableCssModuleTSDeclaration
 
 import EnableCssModuleTSDeclaration from '@zh/shared/config/output/enableCssModuleTSDeclaration.md';
